pyproject.toml
[project]
name = "yolo-uv"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "paho-mqtt>=2.1.0",
    "ultralytics>=8.3.176",
]
